awd_midterm_rest_api/requirements.txt
Django==4.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
coverage==7.2.7
